This is apmonkey's patch from [ 1965920 ] to make routing work in very big midlets, by using ints everywhere instead of shorts.

+++ GpsMidGraph/de/ueller/gpsMid/mapData/RouteTile.java
@@ -118,7 +118,7 @@ public class RouteTile extends RouteBaseTile {
 //			ts.readInt();
 			n.conSize=ts.readByte();
 //			n.fid=fileId;
-			n.id=(short) (i+minId);
+			n.id=i+minId;
 			nodes[i]=n;
 		}
 		ts.close();
